tidy response types
